+++ src/ports/postgres/customer.ts
@@ -1,4 +1,4 @@
-import { Customer } from "@/domain/entities/customer-entity";
+import { Customer } from "@/domain/entities/customer";
 
 export interface CustomerRepositoryPort {
     save(customer: Customer): Promise<Customer>
