+++ src/ports/services/customer.ts
@@ -1,4 +1,4 @@
-import { Customer } from "@/domain/entities/customer-entity";
+import { Customer } from "@/domain/entities/customer";
 
 
 export interface CustomerServicePort {
